% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_translators.R
\name{linear_model_ode}
\alias{linear_model_ode}
\alias{linear_model_ode_threecmt_ev}
\alias{linear_model_ode_threecmt_iv}
\alias{linear_model_ode_twocmt_ev}
\alias{linear_model_ode_twocmt_iv}
\alias{linear_model_ode_onecmt_ev}
\alias{linear_model_ode_onecmt_iv}
\title{Generate linear pharmacokinetic model ordinary differential equation code}
\usage{
linear_model_ode(..., compartments = NULL)

linear_model_ode_threecmt_ev(ka, cl, q, q2, v, vp, vp2, compartments = NULL)

linear_model_ode_threecmt_iv(cl, q, q2, v, vp, vp2, compartments = NULL)

linear_model_ode_twocmt_ev(ka, cl, q, v, vp, compartments = NULL)

linear_model_ode_twocmt_iv(cl, q, v, vp, compartments = NULL)

linear_model_ode_onecmt_ev(ka, cl, v, compartments = NULL)

linear_model_ode_onecmt_iv(cl, v, compartments = NULL)
}
\arguments{
\item{...}{passed to other methods}

\item{compartments}{Character vector of compartment names}

\item{ka}{Absorption rate as an unquoted string}

\item{cl}{Clearanc as an unquoted string}

\item{q}{Intercompartmental clearance between the first and second compartment as an unquoted string}

\item{q2}{Intercompartmental clearance between the first and third compartment as an unquoted string}

\item{v}{Central volume of distribution as an unquoted string}

\item{vp}{First peripheral volume of distribution as an unquoted string}

\item{vp2}{Second peripheral volume of distribution as an unquoted string}
}
\value{
A character vector of code to insert
}
\description{
Generate linear pharmacokinetic model ordinary differential equation code
}
\section{Functions}{
\itemize{
\item \code{linear_model_ode_threecmt_ev}: Three compartment extravascular model

\item \code{linear_model_ode_threecmt_iv}: Three compartment intravascular model

\item \code{linear_model_ode_twocmt_ev}: Two compartment extravascular model

\item \code{linear_model_ode_twocmt_iv}: Two compartment extravascular model

\item \code{linear_model_ode_onecmt_ev}: One compartment extravascular model

\item \code{linear_model_ode_onecmt_iv}: One compartment intravascular model
}}

\examples{
# Manually specify the model to use
linear_model_ode_onecmt_ev(ka=ka_drug1, cl=cl_drug1, v=v_drug1)
# Automatically detect the model to use
linear_model_ode(ka=ka_drug1, cl=cl_drug1, v=v_drug1)
}
